# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "AnalysisOut": {
        "properties": {
          "class_probs": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Class Probs",
            "type": "object"
          },
          "dispute_id": {
            "title": "Dispute Id",
            "type": "string"
          },
          "explanation_bn": {
            "title": "Explanation Bn",
            "type": "string"
          },
          "explanation_en": {
            "title": "Explanation En",
            "type": "string"
          },
          "explanation_sections": {
            "additionalProperties": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            },
            "title": "Explanation Sections",
            "type": "object"
          },
          "flags": {
            "items": {
              "type": "string"
            },
            "title": "Flags",
            "type": "array"
          },
          "injection_detected": {
            "title": "Injection Detected",
            "type": "boolean"
          },
          "model_versions": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Model Versions",
            "type": "object"
          },
          "order_id": {
            "title": "Order Id",
            "type": "string"
          },
          "recommendation": {
            "title": "Recommendation",
            "type": "string"
          },
          "route": {
            "title": "Route",
            "type": "string"
          },
          "route_reasons": {
            "items": {
              "type": "string"
            },
            "title": "Route Reasons",
            "type": "array"
          },
          "timeline": {
            "items": {
              "$ref": "#/components/schemas/TimelineEventOut"
            },
            "title": "Timeline",
            "type": "array"
          }
        },
        "required": [
          "dispute_id",
          "order_id",
          "timeline",
          "class_probs",
          "flags",
          "injection_detected",
          "recommendation",
          "route",
          "route_reasons",
          "explanation_en",
          "explanation_bn",
          "explanation_sections",
          "model_versions"
        ],
        "title": "AnalysisOut",
        "type": "object"
      },
      "AnalystCaseOut": {
        "properties": {
          "analysis": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AnalysisOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "buyer": {
            "$ref": "#/components/schemas/BuyerPublic"
          },
          "decisions": {
            "items": {
              "$ref": "#/components/schemas/DecisionRecordOut"
            },
            "title": "Decisions",
            "type": "array"
          },
          "dispute": {
            "$ref": "#/components/schemas/DisputeOut"
          },
          "order": {
            "$ref": "#/components/schemas/OrderOut"
          },
          "seller": {
            "$ref": "#/components/schemas/SellerPublic"
          },
          "seller_trust": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SnapshotOut"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "dispute",
          "order",
          "buyer",
          "seller",
          "seller_trust",
          "analysis",
          "decisions"
        ],
        "title": "AnalystCaseOut",
        "type": "object"
      },
...
      "BuyerEvidenceRequest": {
        "properties": {
          "evidence_text": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Evidence Text",
            "type": "string"
          }
        },
        "required": [
          "evidence_text"
        ],
        "title": "BuyerEvidenceRequest",
        "type": "object"
      },
      "BuyerPublic": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "wallet_no": {
            "title": "Wallet No",
            "type": "string"
          }
        },
        "required": [
          "id",
          "display_name",
          "wallet_no",
          "created_at"
        ],
        "title": "BuyerPublic",
        "type": "object"
      },
...
      "ClaimType": {
        "enum": [
          "NOT_RECEIVED",
          "WRONG_ITEM",
          "DAMAGED",
          "NOT_AS_DESCRIBED",
          "OTHER"
        ],
        "title": "ClaimType",
        "type": "string"
      },
...
      "CreateDisputeRequest": {
        "properties": {
          "claim_text": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Claim Text",
            "type": "string"
          },
          "claim_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ClaimType"
              },
              {
                "type": "null"
              }
            ]
          },
          "evidence_text": {
            "default": "",
            "maxLength": 4000,
            "title": "Evidence Text",
            "type": "string"
          },
          "order_id": {
            "maxLength": 40,
            "minLength": 1,
            "title": "Order Id",
            "type": "string"
          }
        },
        "required": [
          "order_id",
          "claim_text"
        ],
        "title": "CreateDisputeRequest",
        "type": "object"
      },
...
      "DecisionRecordOut": {
        "properties": {
          "analyst_id": {
            "title": "Analyst Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "decision": {
            "$ref": "#/components/schemas/AnalystDecisionType"
          },
          "note": {
            "title": "Note",
            "type": "string"
          }
        },
        "required": [
          "analyst_id",
          "decision",
          "note",
          "created_at"
        ],
        "title": "DecisionRecordOut",
        "type": "object"
      },
...
      "DisputeOut": {
        "properties": {
          "claim_text": {
            "title": "Claim Text",
            "type": "string"
          },
          "claim_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ClaimType"
              },
              {
                "type": "null"
              }
            ]
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/EvidenceOut"
            },
            "title": "Evidence",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "opened_at": {
            "format": "date-time",
            "title": "Opened At",
            "type": "string"
          },
          "opened_by": {
            "$ref": "#/components/schemas/Party"
          },
          "order_id": {
            "title": "Order Id",
            "type": "string"
          },
          "seller_deadline": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seller Deadline"
          },
          "seller_responded_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seller Responded At"
          },
          "seller_response_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seller Response Text"
          },
          "status": {
            "$ref": "#/components/schemas/DisputeStatus"
          }
        },
        "required": [
          "id",
          "order_id",
          "status",
          "opened_by",
          "claim_text",
          "claim_type",
          "opened_at",
          "seller_deadline",
          "seller_response_text",
          "seller_responded_at",
          "evidence"
        ],
        "title": "DisputeOut",
        "type": "object"
      },
...
      "ErrorBody": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorOut": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorOut",
        "type": "object"
      },
...
      "SellerPublic": {
        "properties": {
          "category": {
            "title": "Category",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "wallet_no": {
            "title": "Wallet No",
            "type": "string"
          }
        },
        "required": [
          "id",
          "display_name",
          "wallet_no",
          "created_at",
          "category"
        ],
        "title": "SellerPublic",
        "type": "object"
      },
      "SellerResponseRequest": {
        "properties": {
          "evidence_text": {
            "default": "",
            "maxLength": 4000,
            "title": "Evidence Text",
            "type": "string"
          },
          "response_text": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Response Text",
            "type": "string"
          }
        },
        "required": [
          "response_text"
        ],
        "title": "SellerResponseRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "SafeOrder",
    "version": "1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/analyst/disputes/{dispute_id}": {
      "get": {
        "description": "Everything an analyst needs: what happened, why it is risky, and what to do next.",
        "operationId": "case_view_api_analyst_disputes__dispute_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "dispute_id",
            "required": true,
            "schema": {
              "title": "Dispute Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalystCaseOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Bad request, for example a wrong delivery code"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Not allowed in the current state"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Request body too large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Validation error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Too many requests"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "A model is not available"
          }
        },
        "summary": "Case View",
        "tags": [
          "analyst"
        ]
      }
    },
...
    "/api/disputes": {
      "post": {
        "description": "The buyer reports a problem. The held money stays in HOLD until an analyst decides.",
        "operationId": "create_dispute_api_disputes_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDisputeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DisputeOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Bad request, for example a wrong delivery code"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Not allowed in the current state"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Request body too large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Validation error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Too many requests"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "A model is not available"
          }
        },
        "summary": "Create Dispute",
        "tags": [
          "disputes"
        ]
      }
    },
    "/api/disputes/{dispute_id}": {
      "get": {
        "description": "The dispute as both parties see it: claim, evidence, deadline and status.\n\nAn addition to the contract. It carries no analysis and no analyst notes; the seller screen\nneeds it to show the claim and the response deadline.",
        "operationId": "get_dispute_api_disputes__dispute_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "dispute_id",
            "required": true,
            "schema": {
              "title": "Dispute Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DisputeOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Bad request, for example a wrong delivery code"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Not allowed in the current state"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Request body too large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Validation error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Too many requests"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "A model is not available"
          }
        },
        "summary": "Get Dispute",
        "tags": [
          "disputes"
        ]
      }
    },
    "/api/disputes/{dispute_id}/analyze": {
      "post": {
        "description": "Run the evidence analyzer. It only suggests; it never moves money.",
        "operationId": "analyze_dispute_api_disputes__dispute_id__analyze_post",
        "parameters": [
          {
            "in": "path",
            "name": "dispute_id",
            "required": true,
            "schema": {
              "title": "Dispute Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Bad request, for example a wrong delivery code"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Not allowed in the current state"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Request body too large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Validation error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Too many requests"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "A model is not available"
          }
        },
        "summary": "Analyze Dispute",
        "tags": [
          "disputes"
        ]
      }
    },
    "/api/disputes/{dispute_id}/buyer-evidence": {
      "post": {
        "description": "Extra evidence from the buyer, for example after an analyst asked for more.\n\nAn addition to the contract: without it a \"request more evidence\" decision would give the\nbuyer no way to answer.",
        "operationId": "buyer_evidence_api_disputes__dispute_id__buyer_evidence_post",
        "parameters": [
          {
            "in": "path",
            "name": "dispute_id",
            "required": true,
            "schema": {
              "title": "Dispute Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BuyerEvidenceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DisputeOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Bad request, for example a wrong delivery code"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Not allowed in the current state"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Request body too large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Validation error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Too many requests"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "A model is not available"
          }
        },
        "summary": "Buyer Evidence",
        "tags": [
          "disputes"
        ]
      }
    },
    "/api/disputes/{dispute_id}/seller-response": {
      "post": {
        "operationId": "seller_response_api_disputes__dispute_id__seller_response_post",
        "parameters": [
          {
            "in": "path",
            "name": "dispute_id",
            "required": true,
            "schema": {
              "title": "Dispute Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SellerResponseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DisputeOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Bad request, for example a wrong delivery code"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Not allowed in the current state"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Request body too large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Validation error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Too many requests"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "A model is not available"
          }
        },
        "summary": "Seller Response",
        "tags": [
          "disputes"
        ]
      }
    },
...
  }
}